pyproject.toml
[project]
name = "linkedin-content-publish"
version = "0.1.0"
description = "Scheduled LinkedIn post publisher via GitHub Actions"
requires-python = ">=3.11"
dependencies = [
    "supabase",
    "requests",
    "python-dotenv",
    "requests-oauthlib>=2.0.0",
    "google-genai",
]
